laravel policy

+++ app/Http/Controllers/PostController.php
@@ -7,13 +7,16 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use Illuminate\Support\Facades\Auth;
 
+use function PHPUnit\Framework\is_null;
+
 class PostController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
-    {
+    {        
+        $this->authorize('viewAny', Post::class);
         $posts = Post::where('user_id', Auth::user()->id)->get();
         return view('resources.post.index', ['posts' => $posts]);
     }
@@ -30,7 +33,7 @@ class PostController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(StorePostRequest $request)
-    {
+    {                
         Post::create([
             'user_id' => Auth::user()->id,
             'subject' => $request->subject,
@@ -38,7 +41,7 @@ class PostController extends Controller
             'status' => ($request->status == "on" ? 1 : 0)
         ]);
 
-        return redirect()->route('post.index')->with('message', 'Post Succesfully Saved!'); 
+        return redirect()->route('post.index')->with('message', 'Post Succesfully Saved!');        
     }
 
     /**
@@ -46,6 +49,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $this->authorize('view', $post);
         return view('resources.post.show', ['post' => $post]);
     }
 
@@ -54,6 +58,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('view', $post);
         return view('resources.post.edit', ['post' => $post]);
     }
 
@@ -61,7 +66,8 @@ class PostController extends Controller
      * Update the specified resource in storage.
      */
     public function update(UpdatePostRequest $request, Post $post)
-    {
+    {                
+        $this->authorize('view', $post);
         $post->update([
             'user_id' => Auth::user()->id,
             'subject' => $request->subject,
@@ -69,7 +75,7 @@ class PostController extends Controller
             'status' => ($request->status == "on" ? 1 : 0)
         ]);
 
-        return redirect()->route('post.index')->with('message', 'Post Succesfully Saved!');  
+        return redirect()->route('post.index')->with('message', 'Post Succesfully Saved!');        
     }
 
     /**
@@ -77,8 +83,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('view', $post);
         $post->delete();
-        return redirect()->route('post.index')->with('message', 'Post Succesfully Deleted!');
+        return redirect()->route('post.index')->with('message', 'Post Succesfully Deleted!');        
     }
 
     public function postIndex() {
